+++ magda-web-client/src/Components/Dataset/Add/DatasetAddMetadataPage.tsx
@@ -33,7 +33,8 @@ import AccrualPeriodicityInput from "./AccrualPeriodicityInput";
 import {
     State,
     saveState,
-    OrganisationAutocompleteChoice
+    OrganisationAutocompleteChoice,
+    createId
 } from "./DatasetAddCommon";
 import DatasetAddPeoplePage from "./Pages/People/DatasetAddPeoplePage";
 import { createPublisher, ensureAspectExists } from "api-clients/RegistryApis";
@@ -69,6 +70,7 @@ import ReviewFilesList from "./ReviewFilesList";
 import ErrorMessageBox from "./ErrorMessageBox";
 
 import helpIcon from "assets/help.svg";
+import { User } from "reducers/userManagementReducer";
 
 const aspects = {
     publishing: datasetPublishingAspect,
@@ -96,6 +98,7 @@ type Props = {
     datasetId: string;
     isNewDataset: boolean;
     history: any;
+    user: User;
 };
 
 class NewDataset extends React.Component<Props, State> {
@@ -119,6 +122,7 @@ class NewDataset extends React.Component<Props, State> {
                 dataset={this.state.dataset}
                 publishing={this.state.datasetPublishing}
                 provenance={this.state.provenance}
+                user={this.props.user}
             />
         ),
         () => (
@@ -498,7 +502,6 @@ class NewDataset extends React.Component<Props, State> {
     async publishDataset() {
         saveState(this.state, this.props.datasetId);
 
-        const id = createId("ds");
         const {
             dataset,
             datasetPublishing,
@@ -544,7 +547,7 @@ class NewDataset extends React.Component<Props, State> {
         });
 
         const inputDataset = {
-            id,
+            id: this.props.datasetId,
             name: dataset.title,
             aspects: {
                 publishing: datasetPublishing,
@@ -564,7 +567,14 @@ class NewDataset extends React.Component<Props, State> {
                     mechanism: provenance.mechanism,
 
                     sourceSystem: provenance.sourceSystem,
-                    derivedFrom: provenance.derivedFrom,
+                    derivedFrom:
+                        provenance.derivedFrom &&
+                        provenance.derivedFrom.map(choice => ({
+                            id: choice.existingId
+                                ? [choice.existingId]
+                                : undefined,
+                            name: !choice.existingId ? choice.name : undefined
+                        })),
                     affiliatedOrganizationIds:
                         provenance.affiliatedOrganizations &&
                         (await Promise.all(
@@ -579,6 +589,11 @@ class NewDataset extends React.Component<Props, State> {
                 }
             }
         };
+
+        if (!inputDataset.aspects["dataset-access-control"].orgUnitOwnerId) {
+            delete inputDataset.aspects["dataset-access-control"];
+        }
+
         await this.props.createRecord(
             inputDataset,
             inputDistributions,
@@ -680,10 +695,6 @@ export default withAddDatasetState(
         )(NewDataset)
     )
 );
-
-function createId(type = "ds") {
-    return `magda-${type}-${uuidv4()}`;
-}
 
 function denormalise(values) {
     const output = {};